requirements.txt
asgiref==3.8.1
aspose-words==24.8.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
contourpy==1.1.1
cycler==0.12.1
Django==4.2.13
django-debug-toolbar==4.4.6
Faker==26.0.0
fonttools==4.53.1
importlib_resources==6.4.0
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.7.5
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
